tx family of programs. Fix bug in 'tx' where not enough precision was used with the FontDict FontMatrix array values, such that a font with a 2048 em square would be converted as having an em-square of 2049. Changes to support the CUBE library element workflow by blending coordinates as absolute values, and with 12 places of precision. Changes to compile 'tx' as a 64 bit program. To change over, simply edit 'FDK/Tools/Programs/public/config/xcconfig/common.xcconfig' to replace ARCHS_STANDARD_32_BIT with ARCHS_STANDARD_64_BIT. Note that this will affect almost all of the FDK programs: I have tested only 'tx' as a 64 bit program.

+++ FDK/Tools/Programs/public/lib/source/absfont/absfont_dump.c
@@ -470,7 +470,7 @@ static void CTL_CDECL dumpInstr(abfGlyphCallbacks *cb, char *fmt, ...)
 			/* Find break point */
 			for (p = &buf[h->left]; *p != ' '; p--)
 				;
-			fprintf(h->fp, "%.*s\n", p - buf, buf);
+			fprintf(h->fp, "%.*s\n", (int)(p - buf), buf);
 			fprintf(h->fp, "%s", p);
 			h->left = 79 - (strlen(p) + 1);
 			}
